Convert base view classes to Typescript
Converted cmmnelementview.js and connector.js to Typescript.
CMMNElementView is now an abstract class on CMMNElementDefinition.
Also made __canHaveAsChild to take a function instead of a string
Also made CanvasElement an abstract class with a generic on joint element.
Moved connector into separate folder, with separate files for classes, and introduced a Coordinates class.